Fix compilation of lldb/lld/clang-extra-tools
clang-extra-tools shall be installed in a very specific directory for it's
automatically compileD.

lldb and lld could be installed in `project/` but they'll be compiled in all 3
stages. We only need them in the last stage.
That's why they must be declared through the `LLVM_ENABLE_PROJECTS` variable,
and placed along side `llvm/` directory (because of how `LLVM_ENABLE_PROJECTS`
work)