Add patch to support external libedit
When libedit isn't installed in non standard directories, CMake files for LLDB
aren't able to find it.

Hence the patch that asks `pkg-config` where libedit is installed.
Given the way sub-project directories are organized, the name of the patch file
says to look for `llvm/../lldb`. Hence the use of `find` to list files in the
`patches/` directory.

TODO: We should be able to not apply the patch if libedit is either not found,
or installed in system directories.